@@ -1,8 +1,6 @@
-package cmd
+package config
 
 import (
-	"github.com/CircleCI-Public/circleci-cli/api"
-
 	"github.com/pkg/errors"
 	"gopkg.in/yaml.v3"
 )
@@ -40,8 +38,7 @@ type processedConfig struct {
 }
 
 // Processes the config down to v2.0, then checks image used against the block list
-func deprecatedImageCheck(response *api.ConfigResponse) error {
-
+func deprecatedImageCheck(response *ConfigResponse) error {
 	aConfig := processedConfig{}
 	err := yaml.Unmarshal([]byte(response.OutputYaml), &aConfig)
 	if err != nil {
